# perf(parser): handle unescaping of escaped markers in parser code directly